Add NvEncode support for some array types (#3)
This PR adds partial supports for the uint64, string, and nvlist array types. Specifically, it adds NvEncode support, so they can be turned into NvLists. It was tested in conjunction with openzfs/zfs#12728, which adds const to a number of the array functions in the libnvpair code so that we don't need to use mutable references everywhere in rust.